+++ src/main/kotlin/wemi/boot/BuildScript.kt
@@ -11,6 +11,7 @@ import wemi.compile.KotlinCompilerFlags
 import wemi.compile.KotlinJVMCompilerFlags
 import wemi.dependency.*
 import wemi.plugin.PluginEnvironment
+import wemi.run.ExitCode
 import wemi.util.*
 import java.io.IOException
 import java.net.URL
@@ -165,7 +166,6 @@ internal fun createProjectFromBuildScriptInfo(buildScriptInfo:BuildScriptInfo?):
                 Keys.internalClasspath set Static(emptyList())
             }
         }
-        Keys.run set Static(0)
 
         locked = true
     }